Split skyportal client from processor
