+++ mirar/processors/avro/__init__.py
@@ -2,4 +2,3 @@
 Module for producing, and broadcasting, avro alerts
 """
 from mirar.processors.avro.ipac_avro_exporter import IPACAvroExporter
-from mirar.processors.avro.send_to_fritz import SendToFritz
