+++ mirar/processors/skyportal/__init__.py
@@ -1,0 +1,4 @@
+"""
+Central module for the skyportal integration
+"""
+from mirar.processors.skyportal.send_to_skyportal import SkyportalSender
